move "save_key_pair" and "get_key_pair" methods from storage service to the key pair service
